Add files via upload
We were given the task to give a short presentation of 5 minutes on a TMDB database. We could mention any anomalies or interesting information that stood out and choose 2 films to analyse and create additional insights on the movies. I decided to find the reason for a movie being so popular based off the inputs given. Additionally the two movies chosen were Minions (most popular) and Alien Zone (Least Popular)